DS: fix tests for quad precision

+++ src/sys/classes/ds/examples/tests/test4.c
@@ -111,7 +111,7 @@ int main(int argc,char **argv)
   /* Eigenvectors */
   j = 2;
   ierr = DSVectors(ds,DS_MAT_X,&j,&rnorm);CHKERRQ(ierr);  /* third eigenvector */
-  ierr = PetscPrintf(PETSC_COMM_WORLD,"Value of rnorm for 3rd vector = %5.3f\n",rnorm);CHKERRQ(ierr);
+  ierr = PetscPrintf(PETSC_COMM_WORLD,"Value of rnorm for 3rd vector = %.3f\n",(double)rnorm);CHKERRQ(ierr);
   ierr = DSVectors(ds,DS_MAT_X,NULL,NULL);CHKERRQ(ierr);  /* all eigenvectors */
   j = 0;
   rnorm = 0.0;
@@ -127,7 +127,7 @@ int main(int argc,char **argv)
   }
   ierr = DSRestoreArray(ds,DS_MAT_X,&X);CHKERRQ(ierr);
   rnorm = PetscSqrtReal(rnorm);
-  ierr = PetscPrintf(PETSC_COMM_WORLD,"Norm of 1st vector = %5.3f\n",rnorm);CHKERRQ(ierr);
+  ierr = PetscPrintf(PETSC_COMM_WORLD,"Norm of 1st vector = %.3f\n",(double)rnorm);CHKERRQ(ierr);
   if (verbose) {
     ierr = PetscPrintf(PETSC_COMM_WORLD,"After vectors - - - - - - - - -\n");CHKERRQ(ierr);
     ierr = DSView(ds,viewer);CHKERRQ(ierr);
